fix launch settings
Note: In launchSettings.json, the "executablePath" value is relative to the "workingDirectory", not to the associated *.csproj file. So if you have relative path components (e.g. '..') in both values (thinking that these are relative to the *.csproj) VS will apply it twice and end up at the wrong path and possibly fail.